@@ -1,42 +1,34 @@
-'use strict';
-
-exports.__esModule = true;
-exports.whyDidYouUpdate = undefined;
-
-var _deepDiff = require('./deepDiff');
-
-var _getDisplayName = require('./getDisplayName');
-
-var _normalizeOptions = require('./normalizeOptions');
-
-var _shouldInclude = require('./shouldInclude');
+import { classifyDiff, DIFF_TYPES } from './deepDiff';
+import { getDisplayName } from './getDisplayName';
+import { normalizeOptions } from './normalizeOptions';
+import { shouldInclude } from './shouldInclude';
 
 function createComponentDidUpdate(opts) {
   return function componentDidUpdate(prevProps, prevState) {
-    var displayName = (0, _getDisplayName.getDisplayName)(this);
+    var displayName = getDisplayName(this);
 
-    if (!(0, _shouldInclude.shouldInclude)(displayName, opts)) {
+    if (!shouldInclude(displayName, opts)) {
       return;
     }
 
-    var propsDiff = (0, _deepDiff.classifyDiff)(prevProps, this.props, displayName + '.props');
-    if (propsDiff.type === _deepDiff.DIFF_TYPES.UNAVOIDABLE) {
+    var propsDiff = classifyDiff(prevProps, this.props, displayName + '.props');
+    if (propsDiff.type === DIFF_TYPES.UNAVOIDABLE) {
       return;
     }
 
-    var stateDiff = (0, _deepDiff.classifyDiff)(prevState, this.state, displayName + '.state');
-    if (stateDiff.type === _deepDiff.DIFF_TYPES.UNAVOIDABLE) {
+    var stateDiff = classifyDiff(prevState, this.state, displayName + '.state');
+    if (stateDiff.type === DIFF_TYPES.UNAVOIDABLE) {
       return;
     }
     opts.notifier(opts.groupByComponent, opts.collapseComponentGroups, displayName, [propsDiff, stateDiff]);
   };
 }
 
-var whyDidYouUpdate = exports.whyDidYouUpdate = function whyDidYouUpdate(React) {
+export var whyDidYouUpdate = function whyDidYouUpdate(React) {
   var opts = arguments.length > 1 && arguments[1] !== undefined ? arguments[1] : {};
 
   var _componentDidUpdate = React.Component.prototype.componentDidUpdate;
-  opts = (0, _normalizeOptions.normalizeOptions)(opts);
+  opts = normalizeOptions(opts);
 
   React.Component.prototype.componentDidUpdate = createComponentDidUpdate(opts);
 
@@ -72,4 +64,4 @@ var whyDidYouUpdate = exports.whyDidYouUpdate = function whyDidYouUpdate(React) 
   return React;
 };
 
-exports.default = whyDidYouUpdate;
+export default whyDidYouUpdate;
