@@ -1,48 +1,46 @@
 'use strict';
 
-Object.defineProperty(exports, '__esModule', {
-  value: true
-});
+exports.__esModule = true;
+exports.classifyDiff = exports.DIFF_TYPES = undefined;
 
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { 'default': obj }; }
+var _pick2 = require('lodash/pick');
 
-var _lodashIsequal = require('lodash.isequal');
+var _pick3 = _interopRequireDefault(_pick2);
 
-var _lodashIsequal2 = _interopRequireDefault(_lodashIsequal);
+var _every2 = require('lodash/every');
 
-var _lodashIsfunction = require('lodash.isfunction');
+var _every3 = _interopRequireDefault(_every2);
 
-var _lodashIsfunction2 = _interopRequireDefault(_lodashIsfunction);
+var _filter2 = require('lodash/filter');
 
-var _lodashKeys = require('lodash.keys');
+var _filter3 = _interopRequireDefault(_filter2);
 
-var _lodashKeys2 = _interopRequireDefault(_lodashKeys);
+var _keys2 = require('lodash/keys');
 
-var _lodashUnion = require('lodash.union');
+var _keys3 = _interopRequireDefault(_keys2);
 
-var _lodashUnion2 = _interopRequireDefault(_lodashUnion);
+var _union2 = require('lodash/union');
 
-var _lodashFilter = require('lodash.filter');
+var _union3 = _interopRequireDefault(_union2);
 
-var _lodashFilter2 = _interopRequireDefault(_lodashFilter);
+var _isFunction2 = require('lodash/isFunction');
 
-var _lodashEvery = require('lodash.every');
+var _isFunction3 = _interopRequireDefault(_isFunction2);
 
-var _lodashEvery2 = _interopRequireDefault(_lodashEvery);
+var _isEqual2 = require('lodash/isEqual');
 
-var _lodashPick = require('lodash.pick');
+var _isEqual3 = _interopRequireDefault(_isEqual2);
 
-var _lodashPick2 = _interopRequireDefault(_lodashPick);
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
-var DIFF_TYPES = {
+var DIFF_TYPES = exports.DIFF_TYPES = {
   UNAVOIDABLE: 'unavoidable',
   SAME: 'same',
   EQUAL: 'equal',
   FUNCTIONS: 'functions'
 };
 
-exports.DIFF_TYPES = DIFF_TYPES;
-var classifyDiff = function classifyDiff(prev, next, name) {
+var classifyDiff = exports.classifyDiff = function classifyDiff(prev, next, name) {
   if (prev === next) {
     return {
       type: DIFF_TYPES.SAME,
@@ -52,7 +50,7 @@ var classifyDiff = function classifyDiff(prev, next, name) {
     };
   }
 
-  if ((0, _lodashIsequal2['default'])(prev, next)) {
+  if ((0, _isEqual3.default)(prev, next)) {
     return {
       type: DIFF_TYPES.EQUAL,
       name: name,
@@ -71,23 +69,23 @@ var classifyDiff = function classifyDiff(prev, next, name) {
   }
 
   var isChanged = function isChanged(key) {
-    return prev[key] !== next[key] && !(0, _lodashIsequal2['default'])(prev[key], next[key]);
+    return prev[key] !== next[key] && !(0, _isEqual3.default)(prev[key], next[key]);
   };
   var isSameFunction = function isSameFunction(key) {
     var prevFn = prev[key];
     var nextFn = next[key];
-    return (0, _lodashIsfunction2['default'])(prevFn) && (0, _lodashIsfunction2['default'])(nextFn) && prevFn.name === nextFn.name;
+    return (0, _isFunction3.default)(prevFn) && (0, _isFunction3.default)(nextFn) && prevFn.name === nextFn.name;
   };
 
-  var keys = (0, _lodashUnion2['default'])((0, _lodashKeys2['default'])(prev), (0, _lodashKeys2['default'])(next));
-  var changedKeys = (0, _lodashFilter2['default'])(keys, isChanged);
+  var keys = (0, _union3.default)((0, _keys3.default)(prev), (0, _keys3.default)(next));
+  var changedKeys = (0, _filter3.default)(keys, isChanged);
 
-  if (changedKeys.length && (0, _lodashEvery2['default'])(changedKeys, isSameFunction)) {
+  if (changedKeys.length && (0, _every3.default)(changedKeys, isSameFunction)) {
     return {
       type: DIFF_TYPES.FUNCTIONS,
       name: name,
-      prev: (0, _lodashPick2['default'])(prev, changedKeys),
-      next: (0, _lodashPick2['default'])(next, changedKeys)
+      prev: (0, _pick3.default)(prev, changedKeys),
+      next: (0, _pick3.default)(next, changedKeys)
     };
   }
 
@@ -98,4 +96,3 @@ var classifyDiff = function classifyDiff(prev, next, name) {
     next: next
   };
 };
-exports.classifyDiff = classifyDiff;
