@@ -1,9 +1,6 @@
 "use strict";
 
-Object.defineProperty(exports, "__esModule", {
-  value: true
-});
-var getDisplayName = function getDisplayName(o) {
+exports.__esModule = true;
+var getDisplayName = exports.getDisplayName = function getDisplayName(o) {
   return o.displayName || o.constructor.displayName || o.constructor.name;
 };
-exports.getDisplayName = getDisplayName;
