@@ -1,23 +1,21 @@
 'use strict';
 
-Object.defineProperty(exports, '__esModule', {
-  value: true
-});
+exports.__esModule = true;
+exports.shouldInclude = undefined;
 
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { 'default': obj }; }
+var _some2 = require('lodash/some');
 
-var _lodashSome = require('lodash.some');
+var _some3 = _interopRequireDefault(_some2);
 
-var _lodashSome2 = _interopRequireDefault(_lodashSome);
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
-var shouldInclude = function shouldInclude(displayName, _ref) {
-  var include = _ref.include;
-  var exclude = _ref.exclude;
+var shouldInclude = exports.shouldInclude = function shouldInclude(displayName, _ref) {
+  var include = _ref.include,
+      exclude = _ref.exclude;
 
-  return (0, _lodashSome2['default'])(include, function (r) {
+  return (0, _some3.default)(include, function (r) {
     return r.test(displayName);
-  }) && !(0, _lodashSome2['default'])(exclude, function (r) {
+  }) && !(0, _some3.default)(exclude, function (r) {
     return r.test(displayName);
   });
 };
-exports.shouldInclude = shouldInclude;
