@@ -2,17 +2,22 @@ var LodashModuleReplacementPlugin = require('lodash-webpack-plugin');
 
 module.exports = {
   type: 'react-component',
-  build: {
-    externals: {
-      react: 'React',
-    },
-    global: 'WhyDidYouUpdate',
-    jsNext: false,
-    umd: true,
+  npm: {
+    umd: {
+      global: 'WhyDidYouUpdate',
+      externals: {
+        'react': 'React'
+      }
+    }
   },
   webpack: {
     extra: {
       plugins: [new LodashModuleReplacementPlugin()],
     },
+  },
+  babel: {
+    plugins: [
+      'babel-plugin-lodash'
+    ]
   }
 };
