@@ -1,11 +1,11 @@
-import _isString from 'lodash.isstring'
+import _ from 'lodash'
 
 import {defaultNotifier} from './defaultNotifier'
 
 export const DEFAULT_INCLUDE = /./
 export const DEFAULT_EXCLUDE = /[^a-zA-Z0-9()]/
 
-const toRegExp = s => _isString(s) ? new RegExp(`^${s}$`) : s
+const toRegExp = s => _.isString(s) ? new RegExp(`^${s}$`) : s
 const toArray = o =>  o ? [].concat(o) : []
 
 export const normalizeOptions = (opts = {}) => {
